% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{calculate_avg_word_length}
\alias{calculate_avg_word_length}
\title{Calculate the average word length in a character vector}
\usage{
calculate_avg_word_length(text)
}
\arguments{
\item{text}{A character vector containing the text to calculate the average word length from}
}
\value{
The average word length as a numeric value
}
\description{
Calculate the average word length in a character vector
}
\examples{
calculate_avg_word_length("The quick brown fox jumps over the lazy dog.")
}
