% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.r
\name{part_of_speech}
\alias{part_of_speech}
\title{Part-of-peech tagging and chunking on text}
\usage{
part_of_speech(text, language = NULL, output = NULL)
}
\arguments{
\item{text}{one element character vector containing the text to analyze (part of
speech). Must be <= 2,000 characters.}

\item{language}{the language \code{text} is in. Value should be \code{NULL} (which will
cause the API to pick the appropriate stemmer) or one of the supported languages
here: \url{http://text-processing.com/docs/tag.html}.}

\item{output}{The tagged (and chunked) text can be returned in one of the following
output formats: "\code{tagged}", "\code{sexpr}" or "\code{iob}". See
\href{http://text-processing.com/docs/tag.html}{the official documentation} for more
information. Use \code{NULL} for default API behavior.}
}
\value{
\code{list} with parsed \code{text}
}
\description{
Not all languages are supported for each \code{output} format. See the
\href{http://text-processing.com/docs/tag.html}{official documentation} for more
information.
}
\note{
The public API is for non-commercial purposes, and each method is throttled to
  1,000 calls per day per IP
}
\examples{
part_of_speech("hello world")

part_of_speech("hello world", output="iob")

# named entity recoginition
part_of_speech("Maine is nice")
}
\references{
\url{http://text-processing.com/docs/tag.html}
}

