% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase.r
\name{phrase_extract}
\alias{phrase_extract}
\title{Phrase extraction & named entity recognition on text}
\usage{
phrase_extract(text, language = c("english", "dutch", "portuguese",
  "spanish"))
}
\arguments{
\item{text}{one element character vector containing the text to analyze (part of
speech). Must be <= 21000 characters.}

\item{language}{the language \code{text} is in. One of "\code{english}",
"\code{dutch}", "\code{portugese}" or "\code{spanish}". here:
\url{http://text-processing.com/docs/phrases.html}.}
}
\value{
\code{list} with parsed \code{text}
}
\description{
Phrase extraction & named entity recognition on text
}
\note{
The public API is for non-commercial purposes, and each method is throttled to
  1,000 calls per day per IP
}
\examples{
phrase_extract("Maine is nice")
}
\references{
\url{http://text-processing.com/docs/phrases.html}
}

