% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimental.r
\name{sentiment}
\alias{sentiment}
\title{Analyze the sentiment of text}
\usage{
sentiment(text, language = c("english", "dutch", "french"))
}
\arguments{
\item{text}{one element character vector containing the text to analyze. . Must be <=
80,000 characters.}

\item{language}{one of "\code{english}", "\code{french}", "\code{dutch}".}
}
\value{
\code{list} containing the \code{label} and \code{probability} estimates
}
\description{
Analyze the sentiment of text
}
\examples{
sentiment("hi friend")
}
\references{
\url{http://text-processing.com/docs/sentiment.html}
}

