% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{sentiment_analysis}
\alias{sentiment_analysis}
\title{Perform sentiment analysis on a text input}
\usage{
sentiment_analysis(text)
}
\arguments{
\item{text}{A character string containing the text to analyze}
}
\value{
A numeric value indicating the overall sentiment of the text
}
\description{
Perform sentiment analysis on a text input
}
\examples{
sentiment_analysis("Hello how are you I'm doing great thanks love")
}
