% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.r
\name{stem}
\alias{stem}
\title{Perform word stemming on text}
\usage{
stem(text, stemmer = NULL, language = NULL)
}
\arguments{
\item{text}{one element character vector containing the text to analyze (stem). Must be
<= 60,000 characters.}

\item{stemmer}{which stemmer to use. Use \code{NULL} to have the API pick an
appropriate stemmer or choose one along with a supported language (see
\url{http://text-processing.com/docs/stem.html}).}

\item{language}{the language \code{text} is in. Value should be \code{NULL} (which will
cause the API to pick the appropriate stemmer) or one of the supported languages
here: \url{http://text-processing.com/docs/stem.html}.}
}
\value{
\code{list} containing the stemmed text
}
\description{
Not all languages are supported by each stemmer. See the
\href{http://text-processing.com/docs/stem.html}{official documentation} for more
information.
}
\note{
The public API is for non-commercial purposes, and each method is throttled to
  1,000 calls per day per IP
}
\examples{
stem("processing")

stem("correr", "snowball", "spanish")
stem("interestingly enough they overwhelmingly supported the wrong candidate")
}
\references{
\url{http://text-processing.com/docs/stem.html}
}

