% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_class.R
\name{text_classification}
\alias{text_classification}
\title{Perform sentiment analysis per sentence on text data}
\usage{
text_classification(text)
}
\arguments{
\item{text}{A character vector containing the text to analyze}
}
\value{
A tibble with the sentiment scores for each sentence
}
\description{
Perform sentiment analysis per sentence on text data
}
\examples{
text_classification("I love this product! It's the best thing I've ever bought.")
}
