% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_process.R
\name{text_preprocessing}
\alias{text_preprocessing}
\title{Preprocess text for analysis}
\usage{
text_preprocessing(text)
}
\arguments{
\item{text}{A character string containing the text to preprocess}
}
\value{
A character vector containing the preprocessed text
}
\description{
Preprocess text for analysis
}
\examples{
text_preprocessing("I love building plastic model kits")
}
